requirements.txt
# Зависимости для Django Rest Framework
Django==5.0.4
djangorestframework==3.14.0
sqlparse==0.5.0
tzdata==2024.1

# Зависимости для FastAPI
fastapi==0.103.1
uvicorn==0.22.0

# Зависимости для Flask-RESTful
Flask==3.0.0
Flask-RESTful==0.3.9
Jinja2==3.1.2
Werkzeug==3.0.0
itsdangerous==2.1.2
click==8.1.3
flask_sqlalchemy==3.1.1


# Зависимости для работы с базами данных (общие)
pydantic==2.9.2
sqlalchemy==2.0.20

# Дополнительные зависимости для работы с изображениями и математическими операциями
numpy==1.26.4
opencv-python-headless==4.9.0.80
pillow==10.3.0

# Типизация и расширения
typing-extensions==4.11.0

